+++ src/app/shared/register/register.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { NgModel, FormGroup, FormBuilder } from '@angular/forms';
 
-import { User } from '../login/user.model';
+import { User } from '../../login/user.model';
 import { AngularFireAuth } from 'angularfire2/auth';
 import { toast } from 'angular2-materialize';
